Replace a href with UI cards
